Enabled leading and trailing whitespace in the output stream
I was running into an issue when decoding fen strings, as it works by
creating a line-buffer which fills until the next game, then flushing
and processing that buffer. In that process, each line is joined into a
string, however if no leading/trailing whitespace is present then some
percentage of fen strings will not be properly formatted (as they are
space delimeted).